Clear quota exceeded flag on new api key, minor refactoring
