+++ src/main/java/ee/carlrobert/codegpt/settings/ServiceSelectionForm.java
@@ -2,7 +2,9 @@ package ee.carlrobert.codegpt.settings;
 
 import com.intellij.openapi.Disposable;
 import com.intellij.openapi.application.ApplicationManager;
+import com.intellij.openapi.editor.colors.EditorColorsManager;
 import com.intellij.openapi.ui.ComboBox;
+import com.intellij.ui.JBColor;
 import com.intellij.ui.TitledSeparator;
 import com.intellij.ui.components.*;
 import com.intellij.util.ui.FormBuilder;
@@ -18,6 +20,7 @@ import ee.carlrobert.codegpt.settings.state.SettingsState;
 import ee.carlrobert.codegpt.completions.you.YouUserManager;
 import ee.carlrobert.codegpt.completions.you.auth.AuthenticationNotifier;
 import ee.carlrobert.codegpt.settings.state.YouSettingsState;
+import ee.carlrobert.codegpt.telemetry.ui.utils.JBLabelUtils;
 import ee.carlrobert.codegpt.util.SwingUtils;
 import ee.carlrobert.llm.client.openai.completion.chat.OpenAIChatCompletionModel;
 import ee.carlrobert.llm.completion.CompletionModel;
@@ -151,12 +154,12 @@ public class ServiceSelectionForm {
             (AuthenticationNotifier) () -> displayWebSearchResultsCheckBox.setEnabled(true));
   }
 
-  public JPanel getForm(SettingsState settings) {
+  public JPanel getForm() {
     var panel = new JPanel(new FlowLayout(FlowLayout.LEADING, 0, 0));
     panel.add(useOpenAIServiceRadioButton);
-    if (settings.isOpenAIQuotaExceeded()) {
-      JBLabel lab1 = new JBLabel("Quota exceeded", JBLabel.LEFT);
-      panel.add(lab1);
+    if (OpenAISettingsState.getInstance().isOpenAIQuotaExceeded()) {
+      panel.add(Box.createHorizontalStrut(4));
+      panel.add(new JBLabel("<html><sup style=\"color: #cc3300;\">quota exceeded</sup></html>"));
     }
     // flow layout's horizontal gap adds annoying horizontal padding on each sides
     panel.add(Box.createHorizontalStrut(16));
