+++ src/main/java/ee/carlrobert/codegpt/settings/SettingsComponent.java
@@ -27,7 +27,7 @@ public class SettingsComponent {
             .resizeX(false)
             .createPanel())
         .addComponent(new TitledSeparator(CodeGPTBundle.get("settingsConfigurable.section.service.title")))
-        .addComponent(serviceSelectionForm.getForm(settings))
+        .addComponent(serviceSelectionForm.getForm())
         .addVerticalGap(8)
         .addComponentFillVertically(new JPanel(), 0)
         .getPanel();
