+++ src/main/resources/messages/codegpt.properties
@@ -1,7 +1,7 @@
 action.editor.group.EditorActionGroup=CodeGPT
 notification.group.name=CodeGPT notification group
 settings.displayName=CodeGPT: Settings
-settingsConfigurable.section.userAuthentication.title=Authentication
+settingsConfigurable.section.userAuthentication.title=Authentication (Optional)
 settingsConfigurable.section.userAuthentication.signIn.label=Sign In
 settingsConfigurable.section.userInformation.title=User Information
 settingsConfigurable.section.integration.apiKeyField.label=API key:
